refactor-#158: change file imports for rest
