+++ src/microservices/booking/rest/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './create-booking.dto';
